modification du type de l'argument amount dans la methode increment pollution dans les class humans et climate

+++ Climate.java
@@ -30,7 +30,7 @@ public class Climate {
         adjustConditions();
     }
 
-    public void incrementPollution(float amount) {
+    public void incrementPollution(double amount) {
         this.pollution += amount;
     }
